requirements.txt
alabaster==0.7.12
asn1crypto==1.3.0
atomicwrites==1.3.0
attrs==19.3.0
azure-common==1.1.25
azure-nspkg==3.0.2
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
Babel==2.8.0
bleach==3.1.4
boto3==1.11.17
botocore==1.14.17
certifi==2020.4.5.1
cffi==1.13.2
chardet==3.0.4
cmarkgfm==0.4.2
colorama==0.4.3
commonmark==0.9.1
cryptography==2.9.2
docutils==0.15.2
entrypoints==0.3
future==0.18.2
idna==2.9
ijson==2.6.1
imagesize==1.2.0
importlib-metadata==1.5.0
Jinja2==2.11.2
jmespath==0.9.5
keyring==21.1.1
MarkupSafe==1.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.2.0
numpy==1.18.1
oscrypto==1.2.0
packaging==20.3
pandas==1.0.3
pkginfo==1.5.0.1
pluggy==0.13.1
pockets==0.9.1
py==1.8.1
pyarrow==0.16.0
pycparser==2.20
pycryptodomex==3.9.7
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyScaffold==3.2.3
pyscaffoldext-markdown==0.3.2
PySocks==1.7.1
pytest==5.4.1
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
readme-renderer==24.0
recommonmark==0.6.0
requests==2.23.0
requests-toolbelt==0.9.1
s3transfer==0.3.3
six==1.14.0
snowballstemmer==2.0.0
snowflake-connector-python==2.2.5
-e git+git@github.com:GEM7318/Snowmobile.git@6b7a9d0bd0f0b1b0abf684230dbb12b3ffe66063#egg=snowmobile
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
tqdm==4.45.0
twine==2.0.0
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.1.0
